# Switch MapBlock compression to zstd