--- a/src/main.cpp
+++ b/src/main.cpp
@@ -113,7 +113,6 @@ static bool determine_subgame(GameParams *game_params);
 static bool run_dedicated_server(const GameParams &game_params, const Settings &cmd_args);
 static bool migrate_map_database(const GameParams &game_params, const Settings &cmd_args);
 static bool recompress_map_database(const GameParams &game_params, const Settings &cmd_args, const Address &addr);
-static bool decompresstimes_database(const GameParams &game_params, const Settings &cmd_args, const Address &addr);
 
 /**********************************************************************/
 
@@ -307,8 +306,6 @@ static void set_allowed_options(OptionList *allowed_options)
 			_("Feature an interactive terminal (Only works when using minetestserver or with --server)"))));
 	allowed_options->insert(std::make_pair("recompress", ValueSpec(VALUETYPE_FLAG,
 			_("Recompress the given database."))));
-	allowed_options->insert(std::make_pair("decompresstimes", ValueSpec(VALUETYPE_FLAG,
-			_("Measure data de-compression times."))));
 #ifndef SERVER
 	allowed_options->insert(std::make_pair("speedtests", ValueSpec(VALUETYPE_FLAG,
 			_("Run speed tests"))));
@@ -895,9 +892,6 @@ static bool run_dedicated_server(const GameParams &game_params, const Settings &
 	if (cmd_args.getFlag("recompress"))
 		return recompress_map_database(game_params, cmd_args, bind_addr);
 
-	if (cmd_args.getFlag("decompresstimes"))
-		return decompresstimes_database(game_params, cmd_args, bind_addr);
-
 	if (cmd_args.exists("terminal")) {
 #if USE_CURSES
 		bool name_ok = true;
@@ -1117,64 +1111,5 @@ static bool recompress_map_database(const GameParams &game_params, const Setting
 	actionstream << "Compression CPU time per block: " << (comptime / count * 1000 * 1000) << "us" << std::endl;
 	actionstream << "Compression speed:              " << (decompdata / comptime / 1024) << " KB/s" << std::endl;
 	actionstream << "############" << std::endl;
-	return true;
-}
-
-static bool decompresstimes_database(const GameParams &game_params, const Settings &cmd_args, const Address &addr)
-{
-	float decomptime = 0.0;
-	u32 compdata = 0;
-
-	Settings world_mt;
-	std::string world_mt_path = game_params.world_path + DIR_DELIM + "world.mt";
-	if (!world_mt.readConfigFile(world_mt_path.c_str())) {
-		errorstream << "Cannot read world.mt!" << std::endl;
-		return false;
-	}
-	std::string backend = world_mt.get("backend");
-	Server server(game_params.world_path, game_params.game_spec, false, addr, false);
-	MapDatabase *db = ServerMap::createDatabase(backend, game_params.world_path, world_mt);
-
-	u32 count = 0;
-	time_t last_update_time = 0;
-	bool &kill = *porting::signal_handler_killstatus();
-
-	// This is ok because the server doesn't actually run
-	std::vector<v3s16> blocks;
-	db->listAllLoadableBlocks(blocks);
-	for (std::vector<v3s16>::const_iterator it = blocks.begin(); it != blocks.end(); ++it) {
-		if (kill) return false;
-
-		std::string data;
-		db->loadBlock(*it, &data);
-		if (data.empty()) {
-			errorstream << "Failed to load block " << PP(*it) << std::endl;
-			return false;
-		}
-
-		std::istringstream iss(data, std::ios_base::binary);
-		compdata += data.size();
-		u8 ver = readU8(iss);
-		MapBlock mb(NULL, *it, &server);
-
-		clock_t t = clock();
-		mb.deSerialize(iss, ver, true);
-		decomptime += (float)(clock() - t) / CLOCKS_PER_SEC;
-
-		if (++count % 0xFF == 0 && time(NULL) - last_update_time >= 2) {
-			std::cerr << (100.0 * count / blocks.size()) << "% completed\r";
-			last_update_time = time(NULL);
-		}
-	}
-	std::cerr << std::endl;
-
-	actionstream << "############" << std::endl;
-	actionstream << "Total blocks:                     " << count << std::endl;
-	actionstream << "Total data (comp.):               " << (compdata / 1024 / 1024) << " MB" << std::endl;
-	actionstream << "Total decompression CPU time:     " << decomptime << "s" << std::endl;
-	actionstream << "Decompression CPU time per block: " << (decomptime / count * CLOCKS_PER_SEC) << "us" << std::endl;
-	actionstream << "Decompression speed:              " << (compdata / decomptime / 1024) << " KB/s" << std::endl;
-	actionstream << "############" << std::endl;
-
 	return true;
 }