… .github/workflows/build.yml
… .github/workflows/macos.yml
… .gitlab-ci.yml
… builtin/settingtypes.txt
… doc/world_format.txt
… src/CMakeLists.txt
… src/mapblock.h
… src/mapgen/mg_schematic.cpp
… src/mapgen/mg_schematic.h
… src/mapnode.cpp
… src/mapnode.h
… src/serialization.cpp
… src/serialization.h
… src/unittest/test_compression.cpp
… util/buildbot/buildwin32.sh
… util/buildbot/buildwin64.sh
… util/ci/common.sh