This add zstd to Minetest.

The code backward, but not forward compatible:
- blocks are compressed with zstd in the wire if the client understands, otherwise zlib is used as before
- all new blocks written to the DB are compressed with zstd
- old blocks are still decompressed with zlib

@sfan5's brotli and @est31's earlier zstd branched helped a lot - although I'm using different zstd APIs.

Questions:
- What about the compression levels? Right now these are zlib specific. For now I simply mapped them to zstd levels 1-10, which really should be good enough. Alternatively I could remove the compression level config, since zstd's default is pretty good.
- Is forward compatibility needed? I.e. should we (somehow) allow old server to still read old maps, even if a newer server has written to them?
- What should the default compression level be? Right now I set it to 0 (which uses the default, which currently is 3). That seems pretty good. Better compression than zlib, and faster too.

**Some stats: (scene with viewing_range of 600, 3700 blocks visible)**
**Zlib-6 (default)**, 54MB DB, 90s to generate (dominated by compression time on save)
**Zlib-3**, 69MB DB, 70s to generate

**Zstd-15**, 50MB, 108s to load
**Zstd-10**, 50MB, 70s to load
**Zstd-6**, 54MB, 67s to load
**Zstd-3 (default)**, 56MB DB, 60s to load
**Zstd-1**, 55MB DB, 60s to load
**Zstd- -1**, 62MB, 60s to load
**Zstd- -10**, 72MB, 60s to load

This shows - as expected - that we get more compression per time from zstd. It also shows that for the tiny things like blocks higher zstd level do not buy much. Perhaps we should default to 1. I'm concluding that zstd below level 3 is no longer the bottelneck. (should confirm with a run with no compression)
For zlib levels other than 3 or 6 do not make much sense, while for zstd level 1 seems to be optimal.

See @ShadowNinja's comment below for more ways to optimize this.

I sporadically watched the map save times. Grew up to 5s for zlib-6, stayed below 2s for zlib-3.
Zstd-10 stayed below 2s, Zstd-1 below 1s.
Similarly Zstd-15 stayed below 5s. Not much improvement there for levels -1 or -10.

~~(I did not test zstd compression levels larger than 10, or negative values - which are faster but with less compression)~~