@lhofhansl 2021-01-05
I can clearly use some help with the build issues. It builds fine on my Linux box.