@lhofhansl 2021-01-05
Looks like `ZSTD_c_compressionLevel` is not defined in older version of zstd, neither ZSTD_e_end, which is need here. Hmmm.