@lhofhansl 2021-01-06
OK. At least it now compiles with the legacy API used...
I have no clue why the unittest fails. It passes on my machine, but I have a newer version of libzstd.