@lhofhansl 2021-01-06
Hah. Older versions of zstd require to initialize a dstream before each decompression. New version do not require that.