@ShadowNinja 2021-01-06
I was actually looking into implementing this myself, but didn't finish it.  Some discussion on other optimizations to the MapBlock format: #3946

Something that might be worth looking into with zstd would be a custom dictionary -- it should improve compression on the small bits that we pass to the compressor.  We compress each part of the mapblock separately, so the inputs are generally 8KiB or less, which isn't much for a compressor to learn from without a precomputed dictionary.  It might also help to compress an entire mapblock instead of compressing nodes, meta, etc separately.  That might also improve performance since you only have to include the compressor startup time once (although I think zstd has some significant optimizations in startup time over zlib).

I wrote a C++ app that converted Minecraft worlds to a Minetest database, profiled it, and found that it spent most of its CPU time in zlib compressing Minetest blocks, so compression performance is definitely important to database performance.