@lhofhansl 2021-01-06
Thanks @ShadowNinja. There's also #10671 (you've seen it). It turns out we write a lot of blocks to disk when the only thing changed is the timestamp.

I've improved startup somewhat by making the stream context thread local.

A pre-computed dictionary is tricky... The data can be only decompressed with that exact same dictionary (unless it is written with the data).

In addition to compression I find that ostringstream is unreasonably slow - especially construction, reducing the number of those will also help a lot.