@lhofhansl 2021-01-06
I do not know how to fix the Android and Windows builds. It seems we're downloading pre-compiled artifacts, I do not have the necessary environments to build one for libzstd.