@lhofhansl 2021-01-07
@ShadowNinja See second commit. For version >= 29 node data and metadata is now compressed together. Can't say I like the code - littered with if-statements, but it works.