@ShadowNinja 2021-01-08
Yeah, I ended up using a `Buffer<u8>` to avoid the overhead of `std::ostringstream`.

I've pushed what I had completed here: https://github.com/ShadowNinja/minetest/commits/mapblock-v29

Only the serialization is completed, and there are a bunch of other changes on top of the compression, but it might be useful.