@lhofhansl 2021-01-08
I like that better. (BTW, as I found out the hard way, you cannot be use the new compress2 zstd API if you want this work with zstd 1.3.3.)
Do you want to continue with your branch, or should I incorporate this in my branch. Fine either way, no ego from my side :)