@lhofhansl 2021-01-08
Re: Returning a buffer from seralizeBulk, won't that call the copy constructor and make another copy of the buffer? (At least on some platforms since NRVO is not guaranteed.)