@lhofhansl 2021-01-09
I removed the node data/metadata combined compression from this patch, since I did not like it.
Perhaps it's best to add zstd first, and then optimize other things around it.