@lhofhansl 2021-01-09
@ShadowNinja new version, that compresses the entire block.