@lhofhansl 2021-01-10
The VS builds at least compile now. Still a linker problem.