@lhofhansl 2021-01-11
The VCPKG builds now are fine. That's all I am going to be able to do, I'll need some help with the buildbot and Android - both of which load some pre-compiled binaries as far as I can see. @sfan5 ?