@lhofhansl 2021-01-21
OK. Docker work too. Now I need to figure out how to get a prebuild zstd image onto minetest.kitsunemimi.pw (and MinGW builds). And the same for Android.