@lhofhansl 2021-01-21
I tested the following scenarios:
- server with zstd loading an old world -> works
- old server trying to load a new world with some blocks written in zstd -> fails, but is expected
- connecting an old client to a new server -> works
- connecting a new client to an old server -> works