@lhofhansl 2021-02-04
In case it's not clear from above... I'm not planning any further work on this until there's a review (and someone points me to how I can get the pre-compiled libraries I need to Windows and Android).