@sfan5 2021-02-04
I plan to review this when time permits but I can answer those questions:
* For win32/win64 I will build a library and upload it to the place where the other libraries are
* For Android a build scripts needs to be added to https://github.com/minetest/minetest_android_deps, then someone needs to run that and push the result to https://github.com/minetest/minetest_android_deps_binaries