@ExeVirus 2021-02-05
Zstd is under BSD license, meaning we could static link it into minetest, meaning you don't need pre-compiled windows or android support. We could even do a .git submodule so we always use the latest release for builds. It's pure C, so I see no issues for any given platform.  Long story short: why are we not compiling this with minetest and keeping it a separate package?