@sfan5 2021-02-05
While that would be convenient we avoid vendoring other libraries too. It can be considered if zstd turns out to be too hard to install on platforms we support, though I doubt that is the case.