@ExeVirus 2021-02-05
Okay, no worries. It was just a suggestion, in that case, the precompiled windows .dlls are here: https://github.com/facebook/zstd/releases/tag/v1.4.8

As for android, lemme look around, I'll update this comment with a link if I find it.