@lhofhansl 2021-02-06
Thanks @ExeVirus . I was thinking about this too, but ultimately think we should avoid that as best as we can. We'd have to track updates ourselves, etc.
Not sure how much we can trust externally built libraries. The zstd ones are probably OK.