@ExeVirus 2021-03-01
Absolutely! What needs to be done? I can help a few hours a week till it's
merged. I know we need to do android deps, what else? Windows testing?

On Sun, Feb 28, 2021, 5:25 PM lhofhansl <notifications@github.com> wrote:

> Time to pick this up again for 5.5?
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/minetest/minetest/pull/10788#issuecomment-787535481>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AKT7N2XQLCFOUDPGFXTYRGDTBK7GBANCNFSM4VWQVMSA>
> .
>