@lhofhansl 2021-03-01
Yeah. Need to build a Android and Windows zstd lib. I do not have environments (and time, frankly :( ) for either.