@lhofhansl 2021-04-09
If we're still considering this, I'll rebase.