@SmallJoker 2021-04-09
Is a hard dependency on zsed required or could this be modular?