@sfan5 2021-04-09
Here's the builds you were waiting for:
* `http://minetest.kitsunemimi.pw/zstd-1.4.9-win32.zip` (can be integrated with buildbot)
* `http://minetest.kitsunemimi.pw/zstd-1.4.9-win64.zip`
* https://github.com/minetest/minetest_android_deps_binaries/pull/5 (probably not testable without android build env)

I'll review the PR itself hopefully soon.
I also think that repeatable benchmarks would be useful here (doesn't need to be in final version). I'm thinking of command line flags that 1) load each map block and time decompression 2) recompress (load + write again) all mapblocks and time compression. You can find inspiration for this in the older issue.