@lhofhansl 2021-04-26
We still have to work out and agree upon the compression levels. Should we use the zstd levels (which have different meanings from the zlib levels)?